Add tidy support for snapshots
This let's us specify exactly which snapshot a given note to update after
snapshot is for.

Closes #2483
